+++ cncmark/README.md
@@ -1,0 +1,3 @@
+# Mark
+
+Mark is a tool for indentifying the vertices and their coordinates of a 3D model. It is used to move the camera to the right position for taking pictures.
