+++ cncmark/tests/fixtures/stl/sample.stl
