--- mark/README.md
@@ -1,3 +1,0 @@
-# Mark
-
-Get lines from 3D models.
